Task included:
#76 Display images for each post
#72 Implement file storages

+++ phase4/app/Http/Controllers/PostController.php
@@ -13,7 +13,9 @@ use Illuminate\Support\Facades\Validator;
 use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Auth;
-
+use Illuminate\Support\Facades\Storage;
+use Illuminate\Filesystem\Filesystem;
+use Illuminate\Support\Collection;
 
 class PostController extends Controller
 {
@@ -41,8 +43,9 @@ class PostController extends Controller
                 'course_num' => 'required|numeric|regex:/[0-9]{3}/',
                 'price' => 'required|numeric|regex:/[0-9]/',
                 'textbookDescription' => 'required|string|max:255',
+                'textbookfile' => 'required',
             ]);
-        } else if ($house_title != NULL) {
+        } /* else if ($house_title != NULL) {
             $request->validate([
                 'house_name' => 'required|string|max:50',
                 'textbook_condition' => 'required',
@@ -50,8 +53,9 @@ class PostController extends Controller
                 'course_num' => 'required|numeric|regex:/[0-9]{3}/',
                 'price' => 'required|numeric|regex:/[0-9]/',
                 'textbookDescription' => 'required|string|max:255',
+                'housingfile' => 'required',
             ]);
-        }
+        } */
 
 
         if ($textbook_title != NULL) {
@@ -60,21 +64,47 @@ class PostController extends Controller
             $course_num = $request->course_num;
             $price = $request->price;
             $description = $request->textbookDescription;
+            $textbookfile = $request->textbookfile;
+            $image = serialize(array());
             DB::table('textbook')->insert([
                 'title' => $textbook_title,
-                'condition' => $textbook_con,
+                'con' => $textbook_con,
                 'course' => $course,
                 'course_num' => $course_num,
                 'price' => $price,
                 'description' => $description,
-                'user_id' => $user_id
+                'user_id' => $user_id,
+                'files' => $image
             ]);
+            
+            $image = unserialize($image);
+            $id = DB::getPdo()->lastInsertId(); 
+            $lastpost = DB::table('textbook')->where('id', $id)->limit(1);
+
+            foreach ($request->textbookfile as $file) {
+                Storage::disk('public')->putFileAs("textbook/" . $id, $file, $file->getClientOriginalName());
+            }
+            $files = Storage::disk('public')->allFiles("textbook/" . $id);
+            
+            foreach ($files as $filepath) {
+                //dd($filepath, Storage::url($filepath));
+                array_push($image, $filepath);
+            }
+            $image = serialize($image);
+            $lastpost->update([
+                'files' => $image
+            ]);
+
+            //$lastpost->save();
+
         } else if ($house_title != NULL) {
             $type = $request->housingType;
             $bedrooms = $request->bedrooms;
             $bathrooms = $request->bathrooms;
             $description = $request->housingDescription;
             $adress = $request->address;
+            $housefile = $request->housefile;
+            $image = serialize(array());
             DB::table('house')->insert([
                 'id' => NULL,
                 'title' => $house_title,
@@ -83,7 +113,26 @@ class PostController extends Controller
                 'bathroom_num' => $bathrooms,
                 'description' => $description,
                 'address' => $adress,
-                'user_id' => $user_id
+                'user_id' => $user_id,
+                'files' => $image
+            ]);
+
+            $image = unserialize($image);
+            $id = DB::getPdo()->lastInsertId(); 
+            $lastpost = DB::table('house')->where('id', $id)->limit(1);
+
+            foreach ($request->housingfile as $file) {
+                Storage::disk('public')->putFileAs("housing/" . $id, $file, $file->getClientOriginalName());
+            }
+            $files = Storage::disk('public')->allFiles("housing/" . $id);
+            
+            foreach ($files as $filepath) {
+                //dd($filepath, Storage::url($filepath));
+                array_push($image, $filepath);
+            }
+            $image = serialize($image);
+            $lastpost->update([
+                'files' => $image
             ]);
 
         } else if ($room_title != NULL) {
@@ -91,13 +140,35 @@ class PostController extends Controller
             $preference = $request->preference;
             $description = $request->roommateDescription;
             $adress = $request->address;
+            $roommatefile = $request->roommatefile;
+            $image = serialize(array());
             DB::table('roommate')->insert([
                 'id' => NULL,
                 'title' => $room_title,
                 'roommate_num' => $roommates,
                 'roommate_gen' => $preference,
                 'description' => $description,
-                'user_id' => $user_id
+                'user_id' => $user_id,
+                'files' => $image
+
+            ]);
+
+            $image = unserialize($image);
+            $id = DB::getPdo()->lastInsertId(); 
+            $lastpost = DB::table('roommate')->where('id', $id)->limit(1);
+
+            foreach ($request->roommatefile as $file) {
+                Storage::disk('public')->putFileAs("roommate/" . $id, $file, $file->getClientOriginalName());
+            }
+            $files = Storage::disk('public')->allFiles("roommate/" . $id);
+            
+            foreach ($files as $filepath) {
+                //dd($filepath, Storage::url($filepath));
+                array_push($image, $filepath);
+            }
+            $image = serialize($image);
+            $lastpost->update([
+                'files' => $image
             ]);
 
         } else {
@@ -106,14 +177,34 @@ class PostController extends Controller
             $sellOption = $request->sellOption;
             $description = $request->genericDescription;
             $adress = $request->address;
+            $genericfile = $request->genericfile;
+            $image = serialize(array());
             DB::table('everything')->insert([
                 'id' => NULL,
                 'title' => $title,
-                'condition' => $con,
+                'con' => $con,
                 'description' => $description,
-                'user_id' => $user_id
+                'user_id' => $user_id,
+                'files' => $image
             ]);
+            
+            $image = unserialize($image);
+            $id = DB::getPdo()->lastInsertId(); 
+            $lastpost = DB::table('everything')->where('id', $id)->limit(1);
 
+            foreach ($request->genericfile as $file) {
+                Storage::disk('public')->putFileAs("generic/" . $id, $file, $file->getClientOriginalName());
+            }
+            $files = Storage::disk('public')->allFiles("generic/" . $id);
+            
+            foreach ($files as $filepath) {
+                //dd($filepath, Storage::url($filepath));
+                array_push($image, $filepath);
+            }
+            $image = serialize($image);
+            $lastpost->update([
+                'files' => $image
+            ]);
 
         }
 
@@ -123,23 +214,45 @@ class PostController extends Controller
 
     public function delete($type, $id)
     {
-
-        //need to add user verification here later
+        // double check the userid
+        $user = Auth::user();
+        $user_id = $user->id;
 
         if ($type == "textbook") {
-            DB::table('textbook')->where('id', $id)->delete();
-            return Redirect::back()->with(['success' => 'You successfully delete a post!']);
+            $item = DB::table('textbook')->where('id', $id)->get();
+            if ($item[0]->user_id == $user_id) {
+                DB::table('textbook')->where('id', $id)->delete();
+                Storage::disk('public')->deleteDirectory('textbook/' . $id);
+                return Redirect::back()->with(['success' => 'You successfully delete a post!']);
+            }
+
         } elseif ($type == "roommate") {
-            DB::table('roommate')->where('id', $id)->delete();
-            return Redirect::back()->with(['success' => 'You successfully delete a post!']);
+            $item = DB::table('roommate')->where('id', $id)->get();
+            if ($item[0]->user_id == $user_id) {
+                DB::table('roommate')->where('id', $id)->delete();
+                Storage::disk('public')->deleteDirectory('roommate/' . $id);
+                return Redirect::back()->with(['success' => 'You successfully delete a post!']);
+            }
+
         } elseif ($type == "house") {
-            DB::table('house')->where('id', $id)->delete();
-            return Redirect::back()->with(['success' => 'You successfully delete a post!']);
+
+            $item = DB::table('house')->where('id', $id)->get();
+            if ($item[0]->user_id == $user_id) {
+                DB::table('house')->where('id', $id)->delete();
+                Storage::disk('public')->deleteDirectory('housing/' . $id);
+                return Redirect::back()->with(['success' => 'You successfully delete a post!']);
+            }
+
         } elseif ($type == "general") {
-            DB::table('everything')->where('id', $id)->delete();
-            return Redirect::back()->with(['success' => 'You successfully delete a post!']);
+            $item = DB::table('everything')->where('id', $id)->get();
+            if ($item[0]->user_id == $user_id) {
+                DB::table('everything')->where('id', $id)->delete();
+                Storage::disk('public')->deleteDirectory('generic/' . $id);
+                return Redirect::back()->with(['success' => 'You successfully delete a post!']);
+            }
+
         }
-        return Redirect::back()->withErrors('Fail to delete the post. Please try again.');
+        return Redirect::back()->withErrors('Fail to delete the post. Please try again later.');
     }
 
     //implement in the future
